pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "mousecoords"
version = "2.0.0"
description = "GUI automation toolkit with computer vision, macro recording, and game automation"
readme = "README.md"
license = "MPL-2.0"
requires-python = ">=3.9"
dependencies = [
    "pyautogui>=0.9",
    "keyboard>=0.13",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
vision = ["opencv-python>=4.5", "numpy>=1.20"]
ocr = ["pytesseract>=0.3", "opencv-python>=4.5", "numpy>=1.20"]
record = ["pynput>=1.7"]
tui = ["rich>=13.0"]
all = [
    "opencv-python>=4.5",
    "numpy>=1.20",
    "pytesseract>=0.3",
    "pynput>=1.7",
    "rich>=13.0",
]

[project.scripts]
mousecoords = "mousecoords.automator:main"

[tool.setuptools.packages.find]
include = ["mousecoords*"]
